--- src/components/footer/footer-button/footer-button.css
@@ -1,23 +1,0 @@
-/* This file contains the styles for the page footer button*/
-
-.footer-button {
-  position: absolute;
-  right: 1%;
-  bottom: 4%;
-  color: var(--color-font);
-  background: var(--current-blue);
-  font-size: 36px;
-  font-family: var(--font-default);
-  height: 56px;
-  width: 56px;
-  border: 10px 10px var(--color-font);
-  border-radius: 50%;
-  box-shadow: 0 0 20px var(--color-shadow);
-  cursor: pointer;
-  transition: transform .2s ease-in-out;
-}
-.footer-button:hover{
-  transform: scale(1.05);
-  filter: invert(85%);
-  box-shadow: none;
-}
