--- src/components/footer/footer-page/footer-page.css
@@ -1,123 +1,0 @@
-/* This file contains the styles for the footer page */
-
-.footer { /* Footer container */
-  position: absolute;
-  right: 3.5vw;
-  bottom: 10vh;
-  width: 200px;
-  height: 250px;
-  text-align: left;
-  background-color: var(--color-background);
-  border: 2px solid var(--current-gray);
-  border-radius: 5px;
-  box-shadow: 1 10px 10px var(--color-shadow);
-  z-index: 10;
-  margin: auto;
-  overflow: auto;
-  overflow-x: hidden;
-  white-space: pre-wrap;
-  font: var(--font-default);
-}
-
-.footer-text { /*  footer text that's not links*/
-  color: var(--color-font-gray);
-  font-family: var(--font-default);
-  cursor: default;
-  left: 0.5vw;
-  position: relative;
-  content: "";
-  display: block;
-  margin: 10px;
-  text-decoration: none;
-}
-.footer-line { /* Footer lines that split content */
-  background-color: var(--current-gray);
-  border: 1px solid var(--current-gray);
-  width: 92%;
-  margin-inline: 4%;
-  border-radius: 1px;
-}
-
-.footer-link { /* individual links */
-  color: var(--color-font-gray);
-  font-family: var(--font-default);
-  cursor: pointer;
-  left: 0.5vw;
-  position: relative;
-  content: "";
-  display: block;
-  margin: 10px;
-  text-decoration: none;
-  transition: transform .2s ease-in-out;
-}
-.footer-link:hover{
-  transform: scale(1.05);
-}
-.footer-link:first-of-type {
-  margin-top: 10px;
-}
-
-.footer-tutorial {
-  position: absolute;
-  right: 3.5vw;
-  bottom: 10vh;
-  width: 15vw;
-  height: 30vh;
-  text-align: left;
-  background-color: var(--color-background);
-  border: 2px solid var(--current-gray);
-  border-radius: 5px;
-  box-shadow: 1 10px 10px var(--color-shadow);
-  z-index: 12;
-  margin: auto;
-  overflow: auto;
-  overflow-x: hidden;
-  font: var(--font-default);
-}
-
-.footer-follow-flex {
-  display: flex;
-  justify-content: center;
-}
-.footer-follow-icon { /* instagram icon next to the text */
-  position: relative;
-  background-color:#121212;
-  width: 14px;
-  height: 14px;
-  padding: 8px;
-  border-radius: 7px;
-  cursor: pointer;
-  margin-inline: 4px;
-}
-.footer-follow-icon img { /* the actual icon */
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  width: 70%;
-  height: 70%;
-  cursor: pointer;
-  filter: var(--svg-invert-white);
-}
-@media (max-width: 1500px) {
-  .footer-tutorial {
-    width: 35vw;
-    height: 25vh;
-    
-  }
-}
-@media (max-width: 1000px) {
-  .footer-tutorial {
-    width: 35vw;
-    height: 25vh;
-    
-  }
-}
-
-@media (max-width: 430px) { /* phone optimization */
-  .footer-tutorial {
-    width: 35vw;
-    height: 40vh;
-    
-  }
-}
