+++ src/components/layout/header.css
@@ -82,6 +82,10 @@
   border: none;
   border-radius: 4px;
   outline: none;
+  transition: background-color 200ms;
+}
+.header-search:hover {
+  background-color: var(--color-hover-input);
 }
 .header-search-input {
   color: var(--color-font);
