--- src/pages/login-stuff/sing-in-header/sign-in-header.css
@@ -1,26 +1,0 @@
-/*styles used for the (very) simple header for the sign in and register page*/
-
-.sign-in-header { /*the header box at the top of the page*/
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 62px;
-    z-index: 1;
-}
-
-.sign-in-header-logo { /*logo formatting used for the header*/
-    position: absolute;
-    left: 80px;
-    top: calc(50% - 1px);
-    transform: translate(0, -50%);
-    height: 35px;
-    cursor: pointer;
-    filter: var(--svg-invert);
-}
-
-@media (max-width: 1160px) { /*logo position is equal to website header logo*/
-    .sign-in-header-logo {
-        left: 10px;
-    }
-}
