+++ src/pages/signup/signup.tsx
@@ -1,5 +1,5 @@
 import Logo from '../../images/memoriter-logo.svg';
-import FooterButton from '../../components/footer/footer-button/footer-button';
+import FooterButton from '../../components/layout/footer/footer';
 import { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { createUserWithEmailAndPassword, getAuth } from 'firebase/auth';
