+++ src/pages/topic-stuff/topic/topic.tsx
@@ -1,9 +1,7 @@
 import './topic.css';
 import { useState, useEffect } from 'react';
-import memoriterLogo from '../../../images/memoriter-logo.svg';
 import BackButton from '../../../components/back-button/BackButton';
 import SettingsIcon from '../../settings/settings-icon/SettingsIcon';
-import FooterButton from '../../../components/footer/footer-button/footer-button';
 import FlashcardForm from '../form/flashcard-form';
 import Backdrop from '../../../components/backdrops/backdrop/backdrop';
 import { Link } from 'react-router-dom';
@@ -332,9 +330,6 @@ function TopicPage() {
           <SettingsIcon />
         </div>
       </main>
-      <footer>
-        <FooterButton />
-      </footer>
     </Layout>
   );
 }
